create fee_config and update_fee_config() to store fee calc info (wip)
